+++ nomad/datasource_nomad_job.go
@@ -132,6 +132,7 @@ func dataSourceJob() *schema.Resource {
 					},
 				},
 			},
+			"update_strategy": updateStrategySchema(),
 			"periodic_config": {
 				Description: "Job Periodic Configuration",
 				Computed:    true,
@@ -204,29 +205,13 @@ func dataSourceJobRead(d *schema.ResourceData, meta interface{}) error {
 	d.Set("stop", job.Stop)
 	d.Set("priority", job.Priority)
 	d.Set("parent_id", job.ParentID)
+
 	d.Set("task_groups", jobTaskGroupsRaw(job.TaskGroups))
 	d.Set("stable", job.Stable)
 	d.Set("all_at_once", job.AllAtOnce)
 	d.Set("constraints", job.Constraints)
-	if job.Periodic != nil {
-		periodic := map[string]interface{}{}
-		if job.Periodic.Enabled != nil {
-			periodic["enabled"] = *job.Periodic.Enabled
-		}
-		if job.Periodic.Spec != nil {
-			periodic["spec"] = *job.Periodic.Spec
-		}
-		if job.Periodic.SpecType != nil {
-			periodic["spec_type"] = *job.Periodic.SpecType
-		}
-		if job.Periodic.ProhibitOverlap != nil {
-			periodic["prohibit_overlap"] = *job.Periodic.ProhibitOverlap
-		}
-		if job.Periodic.TimeZone != nil {
-			periodic["timezone"] = *job.Periodic.TimeZone
-		}
-		d.Set("periodic_config", []map[string]interface{}{periodic})
-	}
+	d.Set("update_strategy", flattenUpdateStrategy(job.Update))
+	d.Set("periodic_config", flattenPeriodicConfig(job.Periodic))
 
 	return nil
 }